refactor: MoveStorage를 MoveStrategy로 패키지 및 클래스명 변경
