+++ src/main/java/janggi/domain/Piece.java
@@ -1,6 +1,6 @@
 package janggi.domain;
 
-import janggi.domain.movestorage.MoveStrategy;
+import janggi.domain.movestrategy.MoveStrategy;
 import janggi.exception.InvalidMoveException;
 
 import java.util.Objects;
