+++ src/main/java/janggi/domain/movestrategy/MaMoveStrategy.java
@@ -1,4 +1,4 @@
-package janggi.domain.movestorage;
+package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
 import janggi.domain.Column;
@@ -8,11 +8,10 @@ import janggi.domain.Row;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SangMoveStrategy implements MoveStrategy {
-    private static final int FORWARD = 3;
-    private static final int DIAGONAL = 2;
+public class MaMoveStrategy implements MoveStrategy {
+    private static final int FORWARD = 2;
+    private static final int DIAGONAL = 1;
 
-    private static final int PATH_STEP_2 = 2;
     private static final int PATH_STEP_1 = 1;
     private static final int PATH_STEP_0 = 0;
 
@@ -26,10 +25,10 @@ public class SangMoveStrategy implements MoveStrategy {
         int diffRow = toRow - fromRow;
         int diffCol = toCol - fromCol;
 
-        boolean isSangMove = (Math.abs(diffRow) == FORWARD && Math.abs(diffCol) == DIAGONAL) ||
+        boolean isMaMove = (Math.abs(diffRow) == FORWARD && Math.abs(diffCol) == DIAGONAL) ||
                 (Math.abs(diffRow) == DIAGONAL && Math.abs(diffCol) == FORWARD);
 
-        if (!isSangMove) {
+        if (!isMaMove) {
             return false;
         }
 
@@ -42,20 +41,12 @@ public class SangMoveStrategy implements MoveStrategy {
             int step1Row = fromRow + (signRow * PATH_STEP_1);
             int step1Col = fromCol + (signCol * PATH_STEP_0);
             movementPathPositions.add(Position.of(Row.of(step1Row), Column.of(step1Col)));
-
-            int step2Row = fromRow + (signRow * PATH_STEP_2);
-            int step2Col = fromCol + (signCol * PATH_STEP_1);
-            movementPathPositions.add(Position.of(Row.of(step2Row), Column.of(step2Col)));
         }
 
         if (Math.abs(diffCol) == FORWARD) {
             int step1Row = fromRow + (signRow * PATH_STEP_0);
             int step1Col = fromCol + (signCol * PATH_STEP_1);
             movementPathPositions.add(Position.of(Row.of(step1Row), Column.of(step1Col)));
-
-            int step2Row = fromRow + (signRow * PATH_STEP_1);
-            int step2Col = fromCol + (signCol * PATH_STEP_2);
-            movementPathPositions.add(Position.of(Row.of(step2Row), Column.of(step2Col)));
         }
 
         for (Position position : movementPathPositions) {
