+++ src/main/java/janggi/domain/movestrategy/PoMoveStrategy.java
@@ -1,11 +1,13 @@
-package janggi.domain.movestorage;
+package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
 import janggi.domain.Column;
+import janggi.domain.Piece;
+import janggi.domain.PieceType;
 import janggi.domain.Position;
 import janggi.domain.Row;
 
-public class ChaMoveStrategy implements MoveStrategy {
+public class PoMoveStrategy implements MoveStrategy {
 
     @Override
     public boolean canMove(Position from, Position to, BoardState boardState) {
@@ -18,6 +20,8 @@ public class ChaMoveStrategy implements MoveStrategy {
             return false;
         }
 
+        int jumpCount = 0;
+
         if (fromRow == toRow) {
             int start = Math.min(fromCol, toCol) + 1;
             int end = Math.max(fromCol, toCol);
@@ -25,7 +29,11 @@ public class ChaMoveStrategy implements MoveStrategy {
             for (int i = start; i < end; i++) {
                 Position position = Position.of(Row.of(fromRow), Column.of(i));
                 if (boardState.hasPieceAt(position)) {
-                    return false;
+                    Piece jumpPiece = boardState.getPieceAt(position);
+                    if (jumpPiece.getPieceType() == PieceType.PO) {
+                        return false;
+                    }
+                    jumpCount++;
                 }
             }
         }
@@ -37,8 +45,23 @@ public class ChaMoveStrategy implements MoveStrategy {
             for (int i = start; i < end; i++) {
                 Position position = Position.of(Row.of(i), Column.of(fromCol));
                 if (boardState.hasPieceAt(position)) {
-                    return false;
+                    Piece jumpPiece = boardState.getPieceAt(position);
+                    if (jumpPiece.getPieceType() == PieceType.PO) {
+                        return false;
+                    }
+                    jumpCount++;
                 }
+            }
+        }
+
+        if (jumpCount != 1) {
+            return false;
+        }
+
+        if (boardState.hasPieceAt(to)) {
+            Piece targetPiece = boardState.getPieceAt(to);
+            if (targetPiece.getPieceType() == PieceType.PO) {
+                return false;
             }
         }
         return true;
