+++ src/test/java/janggi/domain/movestrategy/SangMoveStrategyTest.java
@@ -1,4 +1,4 @@
-package janggi.domain.movestorage;
+package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
 import janggi.domain.Column;
@@ -12,7 +12,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class SangMoveStorageTest {
+class SangMoveStrategyTest {
 
     private static class FakeBoard implements BoardState {
         @Override
